Move to useSelector, and put stuff in its own util
